consolidate the unseal functionality to use a
digest of the wolfBoot text to create the policy that is used to seal the pubkey. tying it to the wolfboot text ensures that wolfBoot isn't modified before releasing the public key. wolfBoot_unseal_at can also be used from a hal to unseal a miscellaneous buffer. this scheme doesn't require an auth key to unseal