Fix wolfBoot_update_trigger sector calculation
wolfBoot_update_trigger is called to initiate an update on the next
boot. It does this by modifying the metadata in the last sector
(selecting one of the two if NVM_FLASH_WRITEONCE is defined).

There's a check to handle the case where the PART_UPDATE_ENDFLAGS isn't
aligned with the WOLFBOOT_SECTOR_SIZE. However, this check was wrong, as
the modulo of these two macros is 0 when aligned, so the check should be
!= 0.