Updates to wolfBoot-tc3xx and test-app ADS projects for ADS 1.10.6
This fix explicitly adds -mtc162 to compiler and linker flags in project
settings to fix a build error when opening projects in ADS 1.10.6.

Opening wolfBoot and test-app projects in ADS 1.10.6 resulted in compilation errors
on tricore asm in IfxCpu.c. Newly created projects in 1.10.6 have -mctc162
passed as a compiler option, but older projects don't. Dropdown menu to
set the CPU architecture is broken in IDE when trying to set in a
project created with older ADS version. Explicitly adding -mtc162 fixes
the issue.